Ajuste de tipo de dado em classes repositorie

+++ udemy/angular_spring_boot/helpdesk/src/main/java/com/aug/helpdesk/repositories/ChamadoRepository.java
@@ -1,11 +1,9 @@
 package com.aug.helpdesk.repositories;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aug.helpdesk.domain.Chamado;
 
-public interface ChamadoRepository extends JpaRepository<Chamado, BigInteger> {
+public interface ChamadoRepository extends JpaRepository<Chamado, Integer> {
 
 }
