+++ udemy/angular_spring_boot/helpdesk/src/main/java/com/aug/helpdesk/repositories/PessoaRepository.java
@@ -1,12 +1,10 @@
 package com.aug.helpdesk.repositories;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aug.helpdesk.domain.Pessoa;
 
-public interface PessoaRepository extends JpaRepository<Pessoa, BigInteger> {
+public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
 
     
 }
